@@ -10,9 +10,9 @@ class EzFlutterMaterialApp extends StatefulWidget {
   List<Locale> locales;
   List<ThemeData> materialThemes;
   bool displayDebugBadge;
-  String initialRoute;
-  Map<String, WidgetBuilder> routes;
-  Locale locale;
+  String? initialRoute;
+  Map<String, WidgetBuilder>? routes;
+  Locale? locale;
 
   EzFlutterMaterialApp(this.app, this.title, this.locales, this.materialThemes,
       this.displayDebugBadge,
@@ -23,8 +23,6 @@ class EzFlutterMaterialApp extends StatefulWidget {
 }
 
 class _EzFlutterMaterialAppState extends State<EzFlutterMaterialApp> {
-  ThemeData currentTheme;
-
   @override
   void initState() {
     EzDynamicTheme().themes = widget.materialThemes;
@@ -37,18 +35,19 @@ class _EzFlutterMaterialAppState extends State<EzFlutterMaterialApp> {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-        initialRoute: widget.initialRoute,
-        routes: widget.routes,
-        locale: widget.locale,
-        title: widget.title,
-        theme: EzDynamicTheme().currentTheme(),
-        localizationsDelegates: [
-          EzTranslationsDelegate(widget.locales),
-          GlobalMaterialLocalizations.delegate,
-          GlobalWidgetsLocalizations.delegate,
-        ],
-        supportedLocales: widget.locales,
-        debugShowCheckedModeBanner: widget.displayDebugBadge,
-        home: widget.app);
+      initialRoute: widget.initialRoute,
+      routes: widget.routes ?? <String, Widget Function(BuildContext)>{},
+      locale: widget.locale,
+      title: widget.title,
+      theme: EzDynamicTheme().currentTheme(),
+      localizationsDelegates: [
+        EzTranslationsDelegate(widget.locales),
+        GlobalMaterialLocalizations.delegate,
+        GlobalWidgetsLocalizations.delegate,
+      ],
+      supportedLocales: widget.locales,
+      debugShowCheckedModeBanner: widget.displayDebugBadge,
+      home: widget.app,
+    );
   }
 }
